Move content types into new Thentos.Ends.Types module and define PNG type
