+++ thentos-core/src/Thentos/Frontend/Types.hs
@@ -13,53 +13,15 @@ import Control.Lens (makeLenses)
 import Control.Monad (mzero)
 import Data.Aeson (FromJSON, ToJSON)
 import Data.ByteString.Builder (toLazyByteString)
-import Data.Proxy (Proxy(Proxy))
-import Data.String.Conversions (ST, LT, SBS, LBS, cs)
+import Data.String.Conversions (ST, SBS, cs)
 import GHC.Generics (Generic)
-import Servant.API (Accept (..), MimeRender (..))
-import Servant.HTML.Blaze (HTML)
-import Text.Blaze.Html (Html, ToMarkup, toHtml)
-import Text.Blaze.Html.Renderer.Pretty (renderHtml)
 import URI.ByteString (RelativeRef, serializeRelativeRef, parseRelativeRef, laxURIParserOptions)
 
 import qualified Data.Aeson as Aeson
 import qualified Generics.Generic.Aeson as Aeson
-import qualified Network.HTTP.Media as Media
 
 import Thentos.Types
 import Thentos.Action.Core (Action)
-
-
--- * content types
-
--- | Html content type with pretty printing.  (See also: package servant-blaze.)
-type HTM = PrettyHTML
-
-renderHTM :: Html -> LBS
-renderHTM = cs . renderHtml
-
-data PrettyHTML
-
-instance Accept PrettyHTML where
-    contentType _ = contentType (Proxy :: Proxy HTML)
-
-instance {-# OVERLAPPABLE #-} ToMarkup a => MimeRender PrettyHTML a where
-    mimeRender _ = renderHTM . toHtml
-
-instance {-# OVERLAPPING #-} MimeRender PrettyHTML Html where
-    mimeRender _ = renderHTM
-
-
-data TextCss
-
-instance Accept TextCss where
-    contentType _ = "text" Media.// "css" Media./: ("charset", "utf-8")
-
-instance MimeRender TextCss LBS    where mimeRender _ = id
-instance MimeRender TextCss SBS    where mimeRender _ = cs
-instance MimeRender TextCss ST     where mimeRender _ = cs
-instance MimeRender TextCss LT     where mimeRender _ = cs
-instance MimeRender TextCss String where mimeRender _ = cs
 
 
 -- * frontend actions
